+++ src/main/java/edu/ie3/datamodel/io/naming/DefaultDirectoryHierarchy.java
@@ -16,7 +16,6 @@ import edu.ie3.datamodel.models.input.connector.Transformer3WInput;
 import edu.ie3.datamodel.models.input.connector.type.LineTypeInput;
 import edu.ie3.datamodel.models.input.connector.type.Transformer2WTypeInput;
 import edu.ie3.datamodel.models.input.connector.type.Transformer3WTypeInput;
-import edu.ie3.datamodel.models.input.graphics.GraphicInput;
 import edu.ie3.datamodel.models.input.system.*;
 import edu.ie3.datamodel.models.input.system.characteristic.WecCharacteristicInput;
 import edu.ie3.datamodel.models.input.system.type.*;
@@ -286,11 +285,7 @@ public class DefaultDirectoryHierarchy implements FileHierarchy {
     THERMAL_RESULTS(
         Constants.RESULT_SUB_TREE.resolve("thermal"),
         false,
-        Stream.of(ThermalUnitResult.class).collect(Collectors.toSet())),
-    GRAPHICS(
-        Constants.INPUT_SUB_TREE.resolve("graphics"),
-        false,
-        Stream.of(GraphicInput.class).collect(Collectors.toSet()));
+        Stream.of(ThermalUnitResult.class).collect(Collectors.toSet()));
     private final Path relPath;
     private final boolean mandatory;
     private final Set<Class<?>> relevantClasses;